ESMCI Version: cime5.4.0-alpha.06 10/25/2017

New user interface features
case.submit now supports arbitrary prerequisites
Add option to define test types in query_testlists
Add ability to cancel batch jobs to the system
Add --baseline-root argument to case.cmpgen_baselines
Use keyword-value pairs for create_clone call

Changes to case control system.
Fix COST_PES and TOTALPES for titan (aprun)
Rename PES_PER_NODE to MAX_MPITASKS_PER_NODE.
Resolve confusion between components and component classes
Add support for python3!
Allow certain batch systems (PBS) to pass flags to case.run/case.test via environment var
Refactor queue selection, use nodemin/nodemax instead of jobmin jobmax
Fix E3SM idmap check.
ERR test rework.
New error trap if invalid "idmap" file is present in seq_maps.rc
Fix generate and compare baseline functions
Add a derived variable NTASKS_PER_INST_COMP where COMP is the component name (ATM, LND, etc)
New batch_env optional entry in config_batch
new batch_cancel field in config_batch
Fix tests for small systems.
Refactor CPLHIST mode and add DATM CPLHIST topo capability
Cleanup of the compvar implementation
Add COMP_ROOT_DIR_ variables so that a component can be moved with the change of a single cime variable
Make sure user changes to wallclock/queue are not lost

New tools
Rewrite of load balancing tool (from CMDV)

Coupler/driver changes
only run driver build namelist after first da cycle

Fixes #1845
Fixes #1426

[BFB]

* jgfouca/branch-for-to-acme-10-25-2017-pr: (437 commits)
  Fixes for 2-case building
  Critical bug fix correcting unicode/str confusion
  Add mpas rpointer fix back in
  Fix config mistakes in config_batch
  Fix archive configuration problem for acme
  Fix R test-opt
  Fix duplication of entries
  Minor config_machines fixup
  Remove extraneous logger.info and change a warning to logger.warning
  Make the prereq test more flexible and correct
  Implement a much simpler script_regression_test tests
  Partial implementation of the prereq test. Still need to find out a generic way of verifying that the dependant job actually depends on the other one
  Revert change from afterok to afterany; this should only be done if the user specifically requests it
  Implements the prereq argument for case.submit, allowing the user to specify jobs which should finish (not necessarily successfully) before running the current job
  Update ChangeLog
  Change case.st_archive dependency string to not include strings indicating unsupported logic. Fix cobalt depend_string. Add depend_separator field to support LSF
  update mpt and pnetcdf on cheyenne
  Update description for TOTALPES
  Include spare nodes in COST_PES
  Update ChangeLog
  ...

+++ scripts/lib/CIME/SystemTests/ers.py
@@ -28,7 +28,7 @@ class ERS(SystemTestsCommon):
         stop_n      = self._case.get_value("STOP_N")
         stop_option = self._case.get_value("STOP_OPTION")
 
-        rest_n = stop_n/2 + 1
+        rest_n = int(stop_n/2 + 1)
         stop_new = stop_n - rest_n
         expect(stop_new > 0, "ERROR: stop_n value {:d} too short {:d} {:d}".format(stop_new,stop_n,rest_n))
